pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "derivative-hedging-rl"
version = "0.3.0"
description = "Adaptive hedging strategies using deep reinforcement learning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "reinforcement-learning",
    "options-trading",
    "hedging",
    "quantitative-finance",
    "deep-learning",
    "pytorch"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core ML/RL
    "torch>=2.0.0",
    "stable-baselines3>=2.1.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    
    # Data Processing & Finance
    "pandas>=2.0.0",
    "yfinance>=0.2.28",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    
    # Hyperparameter Optimization & Visualization
    "optuna>=3.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # Web Framework & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",  # For async PostgreSQL
    
    # Caching & Task Queue
    "redis>=5.0.0",
    "celery>=5.3.0",
    
    # Monitoring & Logging
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.19.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "rich>=13.6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "hypothesis>=6.88.0",
    
    # Code Quality
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pylint>=3.0.0",
    "pre-commit>=3.4.0",
    
    # Documentation
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

hyperparameter-tuning = [
    "optuna>=3.4.0",
    "ray[tune]>=2.7.0",
]

visualization = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
]

dashboard = [
    "streamlit>=1.28.0",
    "streamlit-aggrid>=0.3.4",
]

ml-extras = [
    "wandb>=0.15.0",
    "tensorboard>=2.14.0",
    "huggingface-hub>=0.17.0",
]

all = [
    "derivative-hedging-rl[dev,hyperparameter-tuning,visualization,dashboard,ml-extras]",
]

[project.urls]
Homepage = "https://github.com/yourusername/Derivative_Hedging_RL"
Documentation = "https://derivative-hedging-rl.readthedocs.io"
Repository = "https://github.com/yourusername/Derivative_Hedging_RL"
"Bug Tracker" = "https://github.com/yourusername/Derivative_Hedging_RL/issues"

[project.scripts]
hedge-train = "src.cli:train"
hedge-eval = "src.cli:evaluate"
hedge-serve = "src.cli:serve"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["py.typed"]

# ============================================
# Black - Code Formatter
# ============================================
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

# ============================================
# isort - Import Sorter
# ============================================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/alembic/*"]

# ============================================
# pytest - Testing Framework
# ============================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow-running tests",
    "gpu: Tests requiring GPU",
]

# ============================================
# mypy - Static Type Checker
# ============================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "stable_baselines3.*",
    "gymnasium.*",
    "plotly.*",
    "streamlit.*",
]
ignore_missing_imports = true

# ============================================
# coverage - Code Coverage
# ============================================
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# ============================================
# pylint - Code Linter
# ============================================
[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]
